パンくずリスト生成ロジックをconcernに切り出し、controllerの重複を解消
- AiGenerationsController・RandomWordsController・IdeasControllerの
  3ファイルにほぼ同じ内容で重複していたパンくず生成メソッド群を
  app/controllers/concerns/placeable_breadcrumbs.rbに集約
- 切り出し前に、現状の表示内容を保証する特性テスト
  (spec/requests/breadcrumbs_spec.rb, 7ケース)を追加
- IdeasController: 415行→293行、RandomWordsController: 196行→74行
  （Metrics/ClassLengthの無効化コメントも不要になった）
- 挙動は変更していないことを、自動テスト・RuboCop・ブラウザでの
  全パターン目視確認で検証済み
